+++ vulkano/src/pipeline/graphics_pipeline/builder.rs
@@ -52,6 +52,7 @@ use pipeline::viewport::ViewportsState;
 use VulkanObject;
 use check_errors;
 use descriptor::pipeline_layout::PipelineLayoutDesc;
+use descriptor::pipeline_layout::PipelineLayoutDescTweaks;
 use descriptor::pipeline_layout::PipelineLayoutSuperset;
 use framebuffer::RenderPassSubpassInterface;
 use vk;
@@ -157,7 +158,24 @@ impl<Vdef, Vs, Vss, Tcs, Tcss, Tes, Tess, Gs, Gss, Fs, Fss, Rp>
     pub fn build(self, device: Arc<Device>)
                  -> Result<GraphicsPipeline<Vdef, Box<PipelineLayoutAbstract + Send + Sync>, Rp>,
                            GraphicsPipelineCreationError> {
+        self.with_auto_layout(device, &[])
+    }
+
+    /// Builds the graphics pipeline.
+    ///
+    /// Configures the inferred layout for each descriptor `(set, binding)` in `dynamic_buffers` to accept dynamic
+    /// buffers.
+    pub fn with_auto_layout(self, device: Arc<Device>, dynamic_buffers: &[(usize, usize)])
+                            -> Result<GraphicsPipeline<Vdef, Box<PipelineLayoutAbstract + Send + Sync>, Rp>,
+                                      GraphicsPipelineCreationError>
+    {
         let pipeline_layout;
+
+        fn tweak<T: PipelineLayoutDesc>(desc: T, dynamic_buffers: &[(usize, usize)]) -> PipelineLayoutDescTweaks<T> {
+            let mut desc = PipelineLayoutDescTweaks::new(desc);
+            desc.set_dynamic_buffers(dynamic_buffers.into_iter().cloned());
+            desc
+        }
 
         if let Some(ref tess) = self.tessellation {
             if let Some(ref gs) = self.geometry_shader {
@@ -191,12 +209,14 @@ impl<Vdef, Vs, Vss, Tcs, Tcss, Tes, Tess, Gs, Gss, Fs, Fss, Rp>
                     return Err(GraphicsPipelineCreationError::GeometryFragmentStagesMismatch(err));
                 }
 
-                pipeline_layout = Box::new(self.vertex_shader.as_ref().unwrap().0.layout().clone()
-                    .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
-                    .union(self.tessellation.as_ref().unwrap().tessellation_control_shader.0.layout().clone())    // FIXME: unwrap()
-                    .union(self.tessellation.as_ref().unwrap().tessellation_evaluation_shader.0.layout().clone())    // FIXME: unwrap()
-                    .union(self.geometry_shader.as_ref().unwrap().0.layout().clone())    // FIXME: unwrap()
-                    .build(device.clone()).unwrap()) as Box<_>; // TODO: error
+                pipeline_layout = Box::new(
+                    tweak(self.vertex_shader.as_ref().unwrap().0.layout().clone()
+                          .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
+                          .union(self.tessellation.as_ref().unwrap().tessellation_control_shader.0.layout().clone())    // FIXME: unwrap()
+                          .union(self.tessellation.as_ref().unwrap().tessellation_evaluation_shader.0.layout().clone())    // FIXME: unwrap()
+                          .union(self.geometry_shader.as_ref().unwrap().0.layout().clone()),    // FIXME: unwrap()
+                          dynamic_buffers
+                    ).build(device.clone()).unwrap()) as Box<_>; // TODO: error
 
             } else {
                 if let Err(err) = tess.tessellation_control_shader
@@ -223,11 +243,13 @@ impl<Vdef, Vs, Vss, Tcs, Tcss, Tes, Tess, Gs, Gss, Fs, Fss, Rp>
                     return Err(GraphicsPipelineCreationError::TessEvalFragmentStagesMismatch(err));
                 }
 
-                pipeline_layout = Box::new(self.vertex_shader.as_ref().unwrap().0.layout().clone()
-                    .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
-                    .union(self.tessellation.as_ref().unwrap().tessellation_control_shader.0.layout().clone())    // FIXME: unwrap()
-                    .union(self.tessellation.as_ref().unwrap().tessellation_evaluation_shader.0.layout().clone())    // FIXME: unwrap()
-                    .build(device.clone()).unwrap()) as Box<_>; // TODO: error
+                pipeline_layout = Box::new(
+                    tweak(self.vertex_shader.as_ref().unwrap().0.layout().clone()
+                          .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
+                          .union(self.tessellation.as_ref().unwrap().tessellation_control_shader.0.layout().clone())    // FIXME: unwrap()
+                          .union(self.tessellation.as_ref().unwrap().tessellation_evaluation_shader.0.layout().clone()),    // FIXME: unwrap()
+                          dynamic_buffers
+                    ).build(device.clone()).unwrap()) as Box<_>; // TODO: error
             }
 
         } else {
@@ -249,10 +271,12 @@ impl<Vdef, Vs, Vss, Tcs, Tcss, Tes, Tess, Gs, Gss, Fs, Fss, Rp>
                     return Err(GraphicsPipelineCreationError::GeometryFragmentStagesMismatch(err));
                 }
 
-                pipeline_layout = Box::new(self.vertex_shader.as_ref().unwrap().0.layout().clone()
-                    .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
-                    .union(self.geometry_shader.as_ref().unwrap().0.layout().clone())    // FIXME: unwrap()
-                    .build(device.clone()).unwrap()) as Box<_>; // TODO: error
+                pipeline_layout = Box::new(
+                    tweak(self.vertex_shader.as_ref().unwrap().0.layout().clone()
+                          .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
+                          .union(self.geometry_shader.as_ref().unwrap().0.layout().clone()),    // FIXME: unwrap()
+                          dynamic_buffers
+                    ).build(device.clone()).unwrap()) as Box<_>; // TODO: error
 
             } else {
                 if let Err(err) = self.fragment_shader
@@ -266,15 +290,16 @@ impl<Vdef, Vs, Vss, Tcs, Tcss, Tes, Tess, Gs, Gss, Fs, Fss, Rp>
                 }
 
                 pipeline_layout =
-                    Box::new(self.vertex_shader
-                                 .as_ref()
-                                 .unwrap()
-                                 .0
-                                 .layout()
-                                 .clone()
-                                 .union(self.fragment_shader.as_ref().unwrap().0.layout().clone())
-                                 .build(device.clone())
-                                 .unwrap()) as Box<_>; // TODO: error
+                    Box::new(
+                        tweak(self.vertex_shader
+                              .as_ref()
+                              .unwrap()
+                              .0
+                              .layout()
+                              .clone()
+                              .union(self.fragment_shader.as_ref().unwrap().0.layout().clone()),
+                              dynamic_buffers
+                        ).build(device.clone()).unwrap()) as Box<_>; // TODO: error
             }
         }